[ fix ] Using chainl1 to parse `App`lications
The rule '<Infer> = <Infer> " " <Check>' is left recursive which
causes all sorts of problems. I have learned of chainl1 thanks to
http://stuckinaninfiniteloop.blogspot.com/2011/10/left-recursion-in-parsec.html
and once we start using it, everything works out really well.

Had to generalise its type to
Parser a -> Parser b -> Parser (a -> b -> a) -> Parser a
to handle my use case but other than that it's alright.

Added input.lin with a couple of examples.